// src/api/generated/model/shortAnswerOptions.ts
/**
 * Generated by orval v8.0.1 🍺
 * Do not edit manually.
 * Measuring Parent API
 * API documentation for Measuring Parent application
 * OpenAPI spec version: 1.0.0
 */
import type { AcceptableAnswer } from "./acceptableAnswer";

export interface ShortAnswerOptions {
  acceptableAnswers?: AcceptableAnswer[];
  caseSensitive?: boolean;
  exactMatch?: boolean;
  placeholder?: string;
}